refactor(web): features/sessions を features/login へリネームし login/logout を対称化
login を単独スライスにまとめると logout ボタン利用側へ重い依存が波及するため、
features/sessions（login）と features/logout に分けている。命名の非対称
（sessions と logout）を解消するため features/sessions を features/login にリネームし、
login/logout を対称なスライス対にする。API・URL・routes はリソース名 sessions のまま。
